+++ src/housing/events.ts
@@ -1,9 +1,5 @@
-import { ConditionType } from './conditions';
-import { ActionType } from './actions';
-
-//
-// TODO: COMPLETE AND ENSURE TO MATCH ACTUAL TYPES
-//
+import { ConditionKind } from './conditions';
+import { ActionKind } from './actions';
 
 export enum EventType {
     JOIN = 'JOIN',
@@ -28,133 +24,180 @@ export enum EventType {
 
 export type Event = {
     type: EventType;
-    cancellable: boolean;
-    actions: ActionType[];
-    conditions: ConditionType[];
+    actions: ActionKind[];
+    conditions: ConditionKind[];
 };
+
+const defaultActions = [
+    ActionKind.CONDITIONAL,
+    ActionKind.CHANGE_PLAYER_GROUP,
+    ActionKind.FULL_HEAL,
+    ActionKind.DISPLAY_TITLE,
+    ActionKind.DISPLAY_ACTION_BAR,
+    ActionKind.RESET_INVENTORY,
+    ActionKind.CHANGE_MAX_HEALTH,
+    ActionKind.PARKOUR_CHECKPOINT,
+    ActionKind.GIVE_ITEM,
+    ActionKind.REMOVE_ITEM,
+    ActionKind.SEND_CHAT_MESSAGE,
+    ActionKind.APPLY_POTION_EFFECT,
+    ActionKind.CLEAR_POTION_EFFECTS,
+    ActionKind.GIVE_EXP_LEVELS,
+    ActionKind.CHANGE_PLAYER_STAT,
+    ActionKind.CHANGE_GLOBAL_STAT,
+    ActionKind.TELEPORT_PLAYER,
+    ActionKind.FAIL_PARKOUR,
+    ActionKind.PLAY_SOUND,
+    ActionKind.SET_COMPASS_TARGET,
+    ActionKind.SET_GAME_MODE,
+    ActionKind.CHANGE_HEALTH,
+    ActionKind.CHANGE_HUNGER_LEVEL,
+    ActionKind.RANDOM_ACTION,
+    ActionKind.TRIGGER_FUNCTION,
+    ActionKind.APPLY_INVENTORY_LAYOUT,
+    ActionKind.ENCHANT_HELD_ITEM,
+    ActionKind.PAUSE_EXECUTION,
+    ActionKind.SET_PLAYER_TEAM,
+    ActionKind.CHANGE_TEAM_STAT,
+    ActionKind.DISPLAY_MENU,
+];
+
+const defaultConditions = [
+    ConditionKind.REQUIRED_GROUP,
+    ConditionKind.PLAYER_STAT,
+    ConditionKind.GLOBAL_STAT,
+    ConditionKind.REQUIRED_PERMISSION,
+    ConditionKind.WITHIN_REGION,
+    ConditionKind.HAS_ITEM,
+    ConditionKind.DOING_PARKOUR,
+    ConditionKind.HAS_POTION_EFFECT,
+    ConditionKind.SNEAKING,
+    ConditionKind.FLYING,
+    ConditionKind.HEALTH,
+    ConditionKind.MAX_HEALTH,
+    ConditionKind.HUNGER,
+    ConditionKind.REQUIRED_GAMEMODE,
+    ConditionKind.PLACEHOLDER_NUMBER,
+    ConditionKind.REQUIRED_TEAM,
+    ConditionKind.TEAM_STAT,
+];
 
 export const JoinEvent: Event = {
     type: EventType.JOIN,
-    cancellable: false,
-    actions: [],
-    conditions: [],
+    actions: [...defaultActions],
+    conditions: [...defaultConditions],
 };
 
 export const QuitEvent: Event = {
     type: EventType.QUIT,
-    cancellable: false,
-    actions: [],
-    conditions: [],
+    actions: [
+        ActionKind.CONDITIONAL,
+        ActionKind.CHANGE_PLAYER_STAT,
+        ActionKind.CHANGE_GLOBAL_STAT,
+        ActionKind.RANDOM_ACTION,
+        ActionKind.TRIGGER_FUNCTION,
+        ActionKind.PAUSE_EXECUTION,
+        ActionKind.CHANGE_TEAM_STAT,
+    ],
+    conditions: [...defaultConditions],
 };
 
 export const DeathEvent: Event = {
     type: EventType.DEATH,
-    cancellable: true,
-    actions: [],
-    conditions: [],
+    actions: [...defaultActions, ActionKind.CANCEL_EVENT],
+    conditions: [...defaultConditions],
 };
 
 export const KillEvent: Event = {
     type: EventType.KILL,
-    cancellable: false,
-    actions: [],
-    conditions: [],
+    actions: [...defaultActions],
+    conditions: [...defaultConditions],
 };
 
 export const RespawnEvent: Event = {
     type: EventType.RESPAWN,
-    cancellable: false,
-    actions: [],
-    conditions: [],
+    actions: [...defaultActions],
+    conditions: [...defaultConditions],
 };
 
 export const GroupChangeEvent: Event = {
     type: EventType.GROUP_CHANGE,
-    cancellable: false,
-    actions: [],
-    conditions: [],
+    actions: [...defaultActions],
+    conditions: [...defaultConditions],
 };
 
 export const PvpStateChangeEvent: Event = {
     type: EventType.PVP_STATE_CHANGE,
-    cancellable: false,
-    actions: [],
-    conditions: [],
+    actions: [...defaultActions],
+    conditions: [...defaultConditions, ConditionKind.PVP_ENABLED],
 };
 
 export const FishCaughtEvent: Event = {
     type: EventType.FISH_CAUGHT,
-    cancellable: true,
-    actions: [],
-    conditions: [],
+    actions: [...defaultActions, ActionKind.CANCEL_EVENT],
+    conditions: [...defaultConditions, ConditionKind.FISHING_ENVIRONMENT],
 };
 
 export const PortalUseEvent: Event = {
     type: EventType.PORTAL_USE,
-    cancellable: false,
-    actions: [],
-    conditions: [],
+    actions: [...defaultActions],
+    conditions: [...defaultConditions, ConditionKind.PORTAL_TYPE],
 };
 
 export const DamageEvent: Event = {
     type: EventType.DAMAGE,
-    cancellable: true,
-    actions: [],
-    conditions: [],
+    actions: [...defaultActions, ActionKind.CANCEL_EVENT],
+    conditions: [
+        ...defaultConditions,
+        ConditionKind.DAMAGE_CAUSE,
+        ConditionKind.DAMAGE_AMOUNT,
+    ],
 };
 
 export const BlockBreakEvent: Event = {
     type: EventType.BLOCK_BREAK,
-    cancellable: false,
-    actions: [],
-    conditions: [],
+    actions: [...defaultActions],
+    conditions: [...defaultConditions, ConditionKind.BLOCK_TYPE],
 };
 
 export const ParkourStartEvent: Event = {
     type: EventType.PARKOUR_START,
-    cancellable: false,
-    actions: [],
-    conditions: [],
+    actions: [...defaultActions],
+    conditions: [...defaultConditions],
 };
 
 export const ParkourFinishEvent: Event = {
     type: EventType.PARKOUR_FINISH,
-    cancellable: false,
-    actions: [],
-    conditions: [],
+    actions: [...defaultActions],
+    conditions: [...defaultConditions],
 };
 
 export const DropItemEvent: Event = {
     type: EventType.DROP_ITEM,
-    cancellable: true,
-    actions: [],
-    conditions: [],
+    actions: [...defaultActions, ActionKind.CANCEL_EVENT],
+    conditions: [...defaultConditions, ConditionKind.IS_ITEM],
 };
 
 export const PickupItemEvent: Event = {
     type: EventType.PICKUP_ITEM,
-    cancellable: true,
-    actions: [],
-    conditions: [],
+    actions: [...defaultActions, ActionKind.CANCEL_EVENT],
+    conditions: [...defaultConditions, ConditionKind.IS_ITEM],
 };
 
 export const ChangeHeldItemEvent: Event = {
     type: EventType.CHANGE_HELD_ITEM,
-    cancellable: true,
-    actions: [],
-    conditions: [],
+    actions: [...defaultActions, ActionKind.CANCEL_EVENT],
+    conditions: [...defaultConditions, ConditionKind.IS_ITEM],
 };
 
 export const ToggleSneakEvent: Event = {
     type: EventType.TOGGLE_SNEAK,
-    cancellable: true,
-    actions: [],
-    conditions: [],
+    actions: [...defaultActions, ActionKind.CANCEL_EVENT],
+    conditions: [...defaultConditions],
 };
 
 export const ToggleFlightEvent: Event = {
     type: EventType.TOGGLE_FLIGHT,
-    cancellable: true,
-    actions: [],
-    conditions: [],
+    actions: [...defaultActions, ActionKind.CANCEL_EVENT],
+    conditions: [...defaultConditions],
 };