+++ src/housing/util.ts
@@ -1,17 +1,12 @@
-//
-// TODO: COMPLETE AND ENSURE TO MATCH ACTUAL TYPES
-//
-
 export type Coordinates = {
-    relZ: number;
-    relY: number;
-    relX: number;
     x: number;
     y: number;
     z: number;
     pitch: number;
     yaw: number;
 };
+
+export type Location = 'Spawn' | 'Invoker' | Coordinates; // we don't support "Current Location" due to its unpredictability
 
 export enum StatMode {
     INCREMENT = 'INCREMENT',
@@ -21,8 +16,16 @@ export enum StatMode {
     DIVIDE = 'DIVIDE',
 }
 
-export enum GameMode {
+export enum Gamemode {
     ADVENTURE = 'ADVENTURE',
     SURVIVAL = 'SURVIVAL',
     CREATIVE = 'CREATIVE',
+}
+
+export enum Comparator {
+    LESS_THAN = 'LESS_THAN',
+    LESS_THAN_OR_EQUAL = 'LESS_THAN_OR_EQUAL',
+    EQUAL = 'EQUAL',
+    GREATER_THAN_OR_EQUAL = 'GREATER_THAN_OR_EQUAL',
+    GREATER_THAN = 'GREATER_THAN',
 }